+++ packages/provisioning/src/rest/v1/index.ts
@@ -1,2 +1,3 @@
 export * from './provisioning-domain-api';
 export * from './provisioning-service';
+export * from './rcs';