+++ packages/provisioning/src/rest/v1/rcs/rcs-api.ts
@@ -1,0 +1,13 @@
+import { LazyProvisioningApiClient } from '../provisioning-service';
+import { RcsAccountsApi } from './accounts';
+
+/**
+ * RCS subsection of the Provisioning API.
+ */
+export class RcsApi {
+  public readonly accounts: RcsAccountsApi;
+
+  constructor(lazyClient: LazyProvisioningApiClient) {
+    this.accounts = new RcsAccountsApi(lazyClient);
+  }
+}