+++ packages/provisioning/tests/rest/v1/provisioning-service.test.ts
@@ -1,6 +1,8 @@
 import { RequestPlugin } from '@sinch/sdk-client/src/plugins/core/request-plugin';
 import { ApiTokenRequest, SinchClientParameters } from '@sinch/sdk-client';
 import { ProvisioningService } from '../../../src';
+import { RcsAccountsApi } from '../../../src/rest/v1/rcs/accounts';
+import { RcsApi } from '../../../src/rest/v1/rcs';
 import { WebhooksApi } from '../../../src/rest/v1/webhooks';
 
 describe('Provisioning Service', () => {
@@ -17,7 +19,7 @@ describe('Provisioning Service', () => {
     jest.clearAllMocks();
   });
 
-  it('should initialize the webhooks API', () => {
+  it('should initialize the webhooks and rcs.accounts APIs', () => {
     const params: SinchClientParameters = {
       projectId: 'PROJECT_ID',
       keyId: 'KEY_ID',
@@ -28,6 +30,9 @@ describe('Provisioning Service', () => {
 
     expect(provisioningService.webhooks).toBeInstanceOf(WebhooksApi);
     expect(provisioningService.webhooks.client.apiClientOptions.hostname).toBe(DEFAULT_HOSTNAME);
+    expect(provisioningService.rcs).toBeInstanceOf(RcsApi);
+    expect(provisioningService.rcs.accounts).toBeInstanceOf(RcsAccountsApi);
+    expect(provisioningService.rcs.accounts.client.apiClientOptions.hostname).toBe(DEFAULT_HOSTNAME);
   });
 
   it('should update the API client config', () => {